Add UMD build
