+++ packages/visual-stack-redux/.gitignore
@@ -1,2 +1,3 @@
+/dist
 /node_modules
 /lib
