+++ packages/visual-stack-redux/src/index.js
@@ -1,44 +1,13 @@
-import { lensPath, lensProp, set, assoc, not, over } from 'ramda';
-import { createAction, handleActions } from 'redux-actions';
+import * as MenuBar from './components/MenuBar';
+import * as Modal from './components/Modal';
+import * as Sidebar from './components/Sidebar';
+import reducer from './actions';
 
-const OPEN_DROPDOWN = '@cjdev/visual-stack-redux/OPEN_DROPDOWN';
-const CLOSE_DROPDOWN = '@cjdev/visual-stack-redux/CLOSE_DROPDOWN';
+export {
+  reducer,
+  MenuBar,
+  Modal,
+  Sidebar,
+};
 
-export const openDropdown = createAction(OPEN_DROPDOWN, (menuBarName, dropDownName) => ({ menuBarName, dropDownName }));
-export const closeDropdown = createAction(CLOSE_DROPDOWN, (menuBarName, dropDownName) => ({ menuBarName, dropDownName }));
-
-const OPEN_MODAL = '@cjdev/visual-stack-redux/OPEN_MODAL';
-const CLOSE_MODAL = '@cjdev/visual-stack-redux/CLOSE_MODAL';
-
-export const openModal = createAction(OPEN_MODAL, (component, props) => ({ component, props }));
-export const closeModal = createAction(CLOSE_MODAL);
-
-const emptyModalState = { component: undefined, props: undefined };
-
-const REGISTER_DROPDOWN = '@cjdev/visual-stack-redux/REGISTER_DROPDOWN';
-const TOGGLE_DROPDOWN = '@cjdev/visual-stack-redux/TOGGLE_DROPDOWN';
-
-export const createNavGroupKey = createAction(REGISTER_DROPDOWN);
-export const toggleNavGroup = createAction(TOGGLE_DROPDOWN);
-
-export default handleActions({
-  [OPEN_DROPDOWN]: (state, { payload: { menuBarName, dropDownName } }) =>
-    set(lensPath(['menuBar', menuBarName, dropDownName, 'open']), true, state),
-
-  [CLOSE_DROPDOWN]: (state, { payload: { menuBarName, dropDownName } }) =>
-    set(lensPath(['menuBar', menuBarName, dropDownName, 'open']), false, state),
-
-  [OPEN_MODAL]: (state, { payload }) =>
-    set(lensProp('modal'), payload, state),
-
-  [CLOSE_MODAL]: state =>
-    set(lensProp('modal'), emptyModalState, state),
-
-  [REGISTER_DROPDOWN]: (state, { payload }) =>
-    over(lensProp('navGroupDropdown'), assoc(payload, false), state),
-
-  [TOGGLE_DROPDOWN]: (state, { payload }) =>
-    over(lensPath(['navGroupDropdown', payload]), not, state),
-
-
-}, { menuBar: {}, modal: emptyModalState, navGroupDropdown: {} });
+export * from './actions';
